Revert "Cleanup error messages"
This reverts commit 5bfd56ff7e604f7cdc4c3016380212e12cca7cc9.

+++ backend/src/friendship/friendship.service.ts
@@ -12,19 +12,24 @@ export class FriendshipService {
 	constructor(@InjectRepository(Friendship) private friendshipRepository: Repository<Friendship>, @InjectRepository(User) private userRepository: Repository<User>, @InjectRepository(Member) private memberRepository: Repository<Member>, @InjectRepository(Channel) private channelRepository: Repository<Channel>, private userService: UserService) {}
 
 	// User connect to friendship socket
-	async userConnect(client: any): Promise<any> {
-		client.join((await this.userService.getMyUserData(client.handshake.headers.authorization)).intraName);
+	async userConnect(client: any, server: any): Promise<any> {
+		const USER_DATA = await this.userService.getMyUserData(client.handshake.headers.authorization);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
+		client.join(USER_DATA.intraName);
 	}
 
 	// User send friend request to friendship room
 	async friendshipRoom(client: any, server: any, intraName: string): Promise<any> {
-		const USER_DATA = await this.userService.getMyUserData(client.handshake.headers.authorization);
 		if (intraName === undefined)
-			return server.to(USER_DATA.intraName).emit('friendshipRoom', { "error": "Invalid body - body must include intraName(string)" });
+			return { error: "Invalid body - body must include intraName(string)" };
+		const USER_DATA = await this.userService.getMyUserData(client.handshake.headers.authorization);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
 		client.join(intraName);
 		const FRIENDSHIP = await this.friendshipRepository.findOne({ where: {senderIntraName: USER_DATA.intraName, receiverIntraName: intraName} });
 		if (FRIENDSHIP === null)
-			return server.to(USER_DATA.intraName).emit('friendshipRoom', { "error": "Invalid intraName - friendship does not exist" });
+			return { error: "Friendship does not exist" };
 		server.to(intraName).emit('friendshipRoom', { "intraName": USER_DATA.intraName, "status": FRIENDSHIP.status });
 	}
 
@@ -69,16 +74,18 @@ export class FriendshipService {
 	// Creates a new friendship
 	async newFriendship(accessToken: string, receiverIntraName: string, status: string): Promise<any> {
 		const USER_DATA = await this.userService.getMyUserData(accessToken);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
 		const ERROR = await this.checkJson(USER_DATA.intraName, receiverIntraName, status);
 		if (ERROR)
 			return ERROR;
 		if (status.toUpperCase() == "ACCEPTED")
-			return { error: "Invalid status - friendship status (ACCEPTED) is not supported" }
+			return { error: "Friendship status (ACCEPTED) is not supported - use PATCH method to edit an existing PENDING friendship to ACCEPTED friendship instead" }
 		if ((await this.friendshipRepository.findOne({ where: {senderIntraName: USER_DATA.intraName, receiverIntraName: receiverIntraName} })) !== null || (await this.friendshipRepository.findOne({ where: {senderIntraName: receiverIntraName, receiverIntraName: USER_DATA.intraName} })) !== null)
-			return { error: "Invalid receiverIntraName - friendship already exist" }
+			return { error: "Friendship already exist - use PATCH method to update or DELETE method to delete this existing entry" }
 		const RECEIVER = await this.userRepository.findOne({ where: {intraName: receiverIntraName} });
 		if (RECEIVER === null)
-			return { error: "Invalid receiverIntraName - user does not exist" };
+			return { error: "ReceiverIntraName error - user does not exist" };
 		const NEW_FRIENDSHIP = new Friendship(USER_DATA.intraName, receiverIntraName, status.toUpperCase());
 		await this.friendshipRepository.save(NEW_FRIENDSHIP);
 		return NEW_FRIENDSHIP;
@@ -87,13 +94,15 @@ export class FriendshipService {
 	// Updates a friendship
 	async updateFriendship(accessToken: string, receiverIntraName: string, status: string): Promise<any> {
 		const USER_DATA = await this.userService.getMyUserData(accessToken);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
 		const ERROR = await this.checkJson(USER_DATA.intraName, receiverIntraName, status);
 		if (ERROR)
 			return ERROR;
 		const RECEIVER = await this.friendshipRepository.findOne({ where: {senderIntraName: receiverIntraName, receiverIntraName: USER_DATA.intraName} });
 		if (status.toUpperCase() == "ACCEPTED") {
 			if (RECEIVER === null)
-				return { error: "Invalid receiverIntraName - Friendship does not exist" }
+				return { error: "Friendship does not exist - use POST method to create" }
 			RECEIVER.status = status.toUpperCase();
 			const MY_CHANNEL = await this.channelRepository.findOne({ where: {owner: {intraName: USER_DATA.intraName}} });
 			const MY_MEMBER = await this.memberRepository.findOne({ where: { user: {intraName: USER_DATA.intraName}, channelId: MY_CHANNEL.channelId}})
@@ -101,6 +110,8 @@ export class FriendshipService {
 			if (FRIEND_DATA.error !== undefined)
 				return FRIEND_DATA;
 			const FRIEND_CHANNEL = await this.channelRepository.findOne({ where: {owner: {intraName: receiverIntraName}} });
+			if (FRIEND_CHANNEL === null)
+				return { error: "invalid intraName - intraName does not exist" }
 			const FRIEND_MEMBER = await this.memberRepository.findOne({ where: { user: {intraName: FRIEND_DATA.intraName}, channelId: FRIEND_CHANNEL.channelId}})
 			if (MY_MEMBER === null)
 				await this.memberRepository.save(new Member(USER_DATA, FRIEND_CHANNEL.channelId, true, false, false, new Date().toISOString()));
@@ -112,7 +123,7 @@ export class FriendshipService {
 		if (status.toUpperCase() == "BLOCKED")
 		{
 			if (FRIENDSHIP === null && RECEIVER === null)
-				return { error: "Invalid receiverIntraName - Friendship does not exist" }
+				return { error: "Friendship does not exist - use POST method to create" }
 			if (FRIENDSHIP === null) {
 				FRIENDSHIP.status = status.toUpperCase();
 				await this.friendshipRepository.save(FRIENDSHIP);
@@ -130,12 +141,14 @@ export class FriendshipService {
 	// Deletes a friendship
 	async	deleteFriendship(accessToken: string, receiverIntraName: string): Promise<any> {
 		const USER_DATA = await this.userService.getMyUserData(accessToken);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
 		const ERROR = await this.checkJson(USER_DATA.intraName, receiverIntraName, "ACCEPTED");
 		if (ERROR)
 			return ERROR;
 		const FRIENDSHIP = await this.friendshipRepository.findOne({ where: [{senderIntraName: USER_DATA.intraName, receiverIntraName: receiverIntraName}, {senderIntraName: receiverIntraName, receiverIntraName: USER_DATA.intraName}] });
 		if (FRIENDSHIP === null)
-			return { error: "Invalid receiverIntraName - friendship does not exist" }
+			return { error: "Friendship does not exist - use POST method to create" }
 		if (FRIENDSHIP.senderIntraName === USER_DATA.intraName || (FRIENDSHIP.receiverIntraName === USER_DATA.intraName && FRIENDSHIP.status.toUpperCase() !== "BLOCKED"))
 			await this.friendshipRepository.delete(FRIENDSHIP);
 		return FRIENDSHIP;
@@ -144,6 +157,8 @@ export class FriendshipService {
 	// Returns current friendship with a user
 	async getFriendshipStatus(accessToken: string, receiverIntraName: string): Promise<any> {
 		const USER_DATA = await this.userService.getMyUserData(accessToken);
+		if (USER_DATA.error !== undefined)
+			return { error: USER_DATA.error };
 		const ERROR = await this.checkJson(USER_DATA.intraName, receiverIntraName, "ACCEPTED");
 		if (ERROR)
 			return ERROR;
